==> src/03_vbData_pe.R <==
#!/opt/common/CentOS_7/R/R-4.0.0/lib64/Rscript
## matrix data and seg files from varbin 
## which R is being used ! must be in single-cell-cnv environment
## R.home() == "/opt/common/CentOS_7/R/R-4.0.0/lib64/R" || stop("Wrong environment, run `module load R`")

source("src/myLib.R")

## Collect arguments
args <- commandArgs(trailingOnly = TRUE)
#' for debugging:
#' args <- c("--sample.name=DD4388", "--input.dir=varbin5k/DD4388", "--output.dir=vbData/DD4388/", "--genome=grch38", "--bin.size=5k", "--aligner=bwa", "--nobad=FALSE", "--overwrite.annotations=FALSE")


## Default setting when no arguments passed
if(length(args) < 1) {
  args <- c("--help")
}
 
## Help section
if("--help" %in% args) {
  cat("
      The R/Varbin Tables Script
 
      Arguments:
      --sample.name=bioID                     - character, path to varbin genome reference files
      --input.dir=varbin5k                    - character, path to varbin genome reference files
      --output.dir=vbData                     - character, path to varbin genome reference files
      --bin.size=(5k,20k,50k,100k,120k,200k)  - character, total bins to compile 5k, 20k
      --aligner=bowtie                        - character, bin size in kb for the analysis
      --genome=grch38                         - character, grch37 or grch38
      --nobad=TRUE                            - logical, wether to exclude bins with bad mapping properties
      --overwrite.annotations=FALSE           - logical, weather to overwrite the cell annotations or not
      --help                                  - print this text
 
      Example:
      Rscript src/06_vbData.R --sample.name=WD5816 --input.dir=varbin20k/ --output.dir=vbData/ --bin.size=20k --aligner=bowtie \
           --nobad=TRUE --overwrite.annotations=FALSE
      Rscript src/06_vbData.R --sample.name=WD5816 --input.dir=varbin5k/ --output.dir=vbData/ --bin.size=5k --aligner=bowtie \
           --nobad=TRUE --overwrite.annotations=FALSE
")
  q(save="no")
}


## Parse arguments (we expect the form --arg=value)
parseArgs <- function(x) strsplit(sub("^--", "", x), "=")

argsDF <- as.data.frame(do.call("rbind", parseArgs(args)))
argsL <- as.list(as.character(argsDF$V2))
names(argsL) <- argsDF$V1

## sandardizing to no dashes and 'k' = thousand bins; not kilobases (kb)
argsL$input.dir <- gsub("\\/+$", "", argsL$input.dir)
argsL$output.dir <- gsub("\\/+$", "", argsL$output.dir)
argsL$bin.size <- gsub("kb", "k", argsL$bin.size)

argsL

## Arg1 default

if(! dir.exists(argsL$input.dir)) {
    stop("input directory ", argsL$input.dir, " not found")
    q(save = "no")
}

## Arg2 default -- create output directory if it doesn't exist
if(! dir.exists(argsL$output.dir)) {dir.create(argsL$output.dir, recursive = TRUE)}


## Arg3 default
if(! argsL$bin.size %in% c("200k", "120k", "100k", "50k", "20k", "5k", "1k")) {
    stop("resources for bin.size", argsL$bin.size, "are not available")
    q(save = "no")
}

## Arg5 default
if(is.null(argsL$overwrite.annotations)) {
    argsL$overwrite.annotations <- FALSE
}


## retrieving data into matrices
## Run parameters
sample.name <- argsL$sample.name
inDir  <- file.path(argsL$input.dir)
outDir <- file.path(argsL$output.dir)
bin.size <- argsL$bin.size
aligner <- argsL$aligner
genome <- argsL$genome
nobad <- argsL$nobad
overwrite.annotations <- argsL$overwrite.annotations

## nobad
if(nobad) {
    ext <- ".bwa_seg_nobad.txt"
} else {
    ext <- ".bwa_seg.txt"
}

## load copy number data
cell.files <- list.files(path = inDir,
                         pattern = paste("_", genome, ".", bin.size, ext, sep = ""))
                             

xx <- do.call(rbind, strsplit(cell.files, split = "\\_"))

## reconstruct cell id
if(sample.name == "1000g") {
    cell.list <- xx[,1]
    xx <- cbind(xx, "b0")
    xx[,2] <- "0"
    xx[,3] <- "0"
} else {
    cell.list <- paste(paste(xx[,1], xx[,2], xx[,3], sep = "_"), xx[,4], sep = ".")
}

cellAnnot <- data.frame(cellID = cell.list,
                        plate = xx[,1],
                        subsample = gsub("(.*)_(.*)_(.*)", "\\1_\\2", xx[,1]),
                        sampleID =  gsub("(.*)_(.*)_(.*)", "\\1[PRM]", xx[,1]),
                        bioID =   gsub("(.*)_(.*)_(.*)", "\\1", xx[,1]),
                        subtype = NA,
                        mold.histology = NA,
                        mold.grade = NA,
                        mold.viability = NA,
                        mold.appearance = NA,
                        barcode = xx[,4],
                        sample.type = NA,
                        gate = NA,
                        path.comments = NA,
                        row.names = cell.list)
annotation.file <- file.path(outDir, paste0(sample.name, "_annotations.txt"))

## conditional if exists
if(overwrite.annotations) {
    write.table(cellAnnot,
                file = annotation.file,                       
                row.names = TRUE, quote = FALSE, sep = "\t",
                na = "")
    message("vbData finished succesfully, w/ annotations overwritten")
} else {
    if(!file.exists(annotation.file)) {
    write.table(cellAnnot,
                file = annotation.file,                       
                row.names = TRUE, quote = FALSE, sep = "\t",
                na = "")
    message("vbData finished succesfully, w/ new annotations created")
    } else {
        message("vbData finished succesfully, prev. annotations preserved")
    }
}


## get coordinate data
chh <- read.table(file.path(inDir, cell.files[1]),
                  header = TRUE)[,1:3]

cat("reading copy number varbin data data...\n")
## get seg.quantal data for each cell
vbd <- lapply(cell.files, function(cell) {
    vbseg = read.table(file.path(inDir, cell),
                       header = TRUE)
    vbseg
})
names(vbd) <- cell.list

rm(xx)

#' vbd is a list object containing the *varbin.data.txt
#' names on the list are cellID's
#' also appends 
select_vbd_column <- function(vbd, column, coord) {
    cc <- do.call(cbind, lapply(vbd, function(i) i[, column]))
    if(!is.null(coord)) {
        out <- data.frame(coord, cc)
    } else {
        out <- cc
    }
    return(out)
}

cat("selecting and writing quantals")
## cbs.seg.quantal is the same as seg.quantal
## chose cbs.seg.quantal colum as it's more descritive of the processing
csq.df <- select_vbd_column(vbd = vbd, column = "seg.mean.quantal",
                            coord = chh)
write.table(csq.df,
            file = gzfile(file.path(outDir,
                                    paste(sample.name, "_",
                                          genome, ".", bin.size,
                                          ".varbin.data.txt.gz", sep = ""))),
            sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE)

cat("estimate fga")
data.ploidy <- round(mean(as.matrix(csq.df[,4:ncol(csq.df)]), na.rm = TRUE))

fga.df <- data.frame(
    cellID = names(csq.df)[4:ncol(csq.df)],
    fga = apply(csq.df[,4:ncol(csq.df)], 2, estimate_fga,
                base.ploidy = data.ploidy),
    mad = apply(csq.df[,4:ncol(csq.df)], 2, mad)
)

write.table(fga.df,
            file = file.path(outDir,
                             paste0(sample.name, "_",
                                   genome, ".", bin.size, ".FGA.txt")),
            sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE)

rm(csq.df)

## seg.mean.LOWESS is the same as cbs.seg;
## using seg.mean.LOWESS as it is more descriptive of column content
## get lowess ratio (lowratio) data (segmented)
cbd.df <- select_vbd_column(vbd, column = "lowess.ratio.quantal",
                            coord = chh)
write.table(cbd.df,
            file = gzfile(file.path(outDir,
                             paste(sample.name, "_", genome, ".", bin.size,
                                   ".varbin.lowess.data.txt.gz", sep = ""))),
            sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE)
rm(cbd.df)

## BIN COUNT DATA: (raw counts + gc-corrected, and unsegmented!)
## cat("reading raw bin counts...\n")
rbc.df <- select_vbd_column(vbd = vbd, column = "count",
                            coord = chh)
write.table(rbc.df,
            file = gzfile(file.path(outDir,
                                    paste(sample.name, "_", genome, ".", bin.size,
                                          ".bin_counts.txt.gz", sep = ""))),
            sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE)
rm(rbc.df)

## MAPD
cat("estimating MAPD...\n")
## get lowess ration (lowratio) data (not segmented)
lbd <- select_vbd_column(vbd = vbd, column = "lowess.ratio",
                         coord = NULL)
## calculate mapd
mapd.qc <- t(apply(lbd, 2, mapd))
mapd.qc <- data.frame(cellID = rownames(mapd.qc), mapd.qc)

## write out
write.table(mapd.qc,
            file = file.path(outDir,
                             paste(sample.name, "_", genome, ".", bin.size,
                                   ".varbin.mapd.qc.txt", sep = "")),
            sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE)

rm(lbd)

## PLOIDY
cat("reading ploidy and stats data...\n")
## write out ploidy
q.ploidy.files <- list.files(inDir, pattern = "quantal.ploidy.txt$",
                            full.names = TRUE)

xx <- do.call(rbind, strsplit(unlist(sapply(strsplit(q.ploidy.files, "/"), `[`, 3)), "_"))

if(sample.name == "1000g") {
    cell.list.qs <- xx[,1]
    xx[, 2] <-  "0"
    xx <- cbind(xx, "0", "b0")
} else {
    cell.list.qs <- paste(paste(xx[,1], xx[,2], xx[,3], sep = "_"), xx[,4], sep = ".")
}
    
quantal.ploidy <- do.call(rbind, lapply(q.ploidy.files, read.delim, nrow = 1))
quantal.ploidy$cellID  <- cell.list.qs

write.table(quantal.ploidy,
            file = file.path(outDir,
                             paste(sample.name, "_", genome, ".", bin.size,
                                   ".quantal.ploidy.txt", sep = "")),
            sep = "\t", quote = FALSE , row.names = FALSE)
rm(quantal.ploidy, xx, cell.list.qs)

## get varbin stats
cat("reading varbin stats data...\n")
## write out cell annotations file

read_varbin_pe_stats <- function(stats.file) {
    cell.id <- gsub(".*/(.*)_(b[0-9]+).*", "\\1.\\2", stats.file )
    ##
    nx <- length(readLines(stats.file))
    if(nx) {
    ss <- read.table(stats.file, skip = 1, header = FALSE,
                     fill = TRUE, row.names = 1)
    sst <- data.frame(t(ss[,2]))
    names(sst) <- paste(rownames(ss), "Reads", sep = ".")
    rownames(sst) = cell.id
    } else {
        sst <- data.frame(Total.Reads = 0, Passed.Reads = 0,
                          Filtered.Reads = 0, Counted.Reads = 0,
                          row.names = cell.id)
    }
    return(sst)
}

vb.stats.files <- list.files(inDir, pattern = ".counts.stats.bed$",
                             full.names = TRUE)

varbin.stats <- do.call(rbind, lapply(vb.stats.files, read_varbin_pe_stats))
varbin.stats <- cbind(cellID = rownames(varbin.stats), varbin.stats)

write.table(varbin.stats, file = file.path(outDir, paste(sample.name, "_", genome, ".varbin.stats.txt", sep = "")),
            sep = "\t", quote = FALSE , row.names = FALSE)

## READ WRITE SEG FILES
cat("reading segment data...\n")
## write out .seg file
seg.files <- list.files(inDir, pattern = paste0("_", genome, ".", bin.size, ".bwa_short_seg.txt$"),
                        full.names = TRUE)

igv.seg <- do.call(rbind, lapply(seg.files, read.delim)) 
igv.seg$ID = gsub("varbin[0-9]+k..(.*)\\.(b[0-9]+)\\.(b[0-9]+).*", "\\1.\\2.\\3", igv.seg$ID)
igv.seg$chrom = gsub("[pq]", "", igv.seg$chrom)
## select(-start, -end, -num)
igv.seg <- igv.seg[, !names(igv.seg) %in% c("start", "end", "num")]
write.table(igv.seg,
            file = file.path(outDir,
                             paste(sample.name, "_", genome, ".", bin.size,
                                   ".varbin.short.cbs.seg", sep = "")),
            sep = "\t", quote = FALSE, row.names = FALSE, col.names = FALSE)


# Extract gc.content from the first file
#% gc.content <- select_vbd_column(vbd = vbd, column = "gc.content")
## scale to geometric mean
#% TARGET_SCALE <- 2^mean(log2(colSums(rbc.df[,-c(1:3)]]+1)))
#% ## run lowess/loess normalization
#% cat("gc corrected bin counts...\n")
#% gc.norm.bin.count <- apply(rbc.df[, -c(1:3)], 2, function(rbc) {
#%     ## estimate bin weights based on gc (4 sig. digits)
#%     wts <- tapply(rbc, round(gc.content, digits = 4), sum)
#%     wts <- wts[as.character(round(gc.content, 4))]
#%     ## log read count
#%     ## TARGET_SCALE <- 1e6
#%     scl <- TARGET_SCALE / sum(rbc +1)
#%     logtn <- log2((rbc + 1) * scl)
#%     ## weighed loess regression
#%     gcb <- loess(logtn ~ gc.content, weights = wts, span = 0.3,
#%                  control = loess.control(surface = "direct"), 
#%                  degree = 2)
#%     ## scaling read bin counts
#%     res <- 2^-gcb$fitted * rbc
#%     res[is.na(res)] <- 0
#%     ## return res
#%     return(res)
#% })
#% 
#% write.table(gc.norm.bin.count,
#%             file = file.path(outDir, paste(sample.name, "_", genome, ".", bin.size,
#%                                            ".gc_corrected.bin_counts.txt",
#%                                            sep = "")),
#%             sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE)
#% 

## __EOF__
